Support versions with multiple "-" or "+" signs
Currently if the version looks like `1.0.0-alpha-beta` it will be treated like it has a `PreRelease` part equal to `alpha`.

The `PreRelease` part needs to be `alpha-beta` if you look at [semver spec](https://semver.org/#spec-item-9), that allows extra dash characters in the `PreRelease` part.

The `BuildMetadata` part has a similar issue, with the plus sign., i. e. for `1.0.0+123+321` it will be computed to `123` and not `123+321`. Multiple pluses in version is not valid semver, and will probably cause an error when used with nuget. Despite that, it still feels like a similar issue should be fixed there.

In sum, for version ``1.0.0-alpha-beta+123+321` this commit will set `MinVerPreRelease` to `alpha-beta` instead of current `alpha` and will set `MinVerBuildMetadata` to `123+321` instead of current `123`.